+++ src/apiflask/openapi.py
@@ -5,6 +5,7 @@ import typing as t
 from apispec import APISpec
 
 from .exceptions import _bad_schema_message
+from .schemas import FileSchema
 from .security import HTTPBasicAuth
 from .security import HTTPTokenAuth
 from .types import HTTPAuthType
@@ -178,6 +179,8 @@ def add_response(
     """
     operation['responses'][status_code] = {}
     if status_code != '204':
+        if isinstance(schema, FileSchema):
+            schema = {'type': schema.type, 'format': schema.format}
         operation['responses'][status_code]['content'] = {
             content_type: {
                 'schema': schema